🚚Move photo page to pages
