+++ src/components/pages/PhotoPage/CategoryField/components/Category.js
