+++ src/components/pages/PhotoPage/CategoryField/components/CategoryDropdown/FieldLabel/components/FieldLabel.scss
