+++ src/components/pages/PhotoPage/CategoryField/components/CategoryDropdown/FieldLabel/components/FieldLabelWithInput.scss
