+++ src/components/pages/PhotoPage/CategoryField/components/CategoryDropdown/FieldLabel/index.js
