+++ src/components/pages/PhotoPage/CategoryField/components/CategoryDropdown/components/CategoryDropdown.js
