+++ src/components/pages/PhotoPage/CategoryField/components/CategoryDropdown/consts/index.js
