+++ src/components/pages/PhotoPage/CategoryField/components/CategoryDropdown/index.js
