+++ src/components/pages/PhotoPage/CategoryField/components/validation/index.js
