+++ src/components/pages/PhotoPage/CategoryField/index.js
