+++ src/components/pages/PhotoPage/Dialogs/Dialogs.scss
