+++ src/components/pages/PhotoPage/index.js
@@ -8,12 +8,12 @@ import Button from "@material-ui/core/Button";
 import Link from "@material-ui/core/Link";
 import { withStyles } from "@material-ui/core/styles";
 
-import config from "../../custom/config";
-import { gtagEvent } from "../../gtag.js";
-import dbFirebase from "../../dbFirebase";
-import { isIphoneWithNotchAndCordova, device } from "../../utils";
+import config from "../../../custom/config";
+import { gtagEvent } from "../../../gtag.js";
+import dbFirebase from "../../../dbFirebase";
+import { isIphoneWithNotchAndCordova, device } from "../../../utils";
 
-import PageWrapper from "../PageWrapper";
+import PageWrapper from "../../PageWrapper";
 import Fields from "./Fields";
 import Dialogs from "./Dialogs";
 import "./style.scss";
