+++ src/components/pages/PhotoPage/style.scss
